requirements.txt
pytest~=9.0.3
requests~=2.33.0
faker~=40.4.0
python-dotenv~=1.2.2